+++ core-services/prompt-registry/src/main/java/com/sap/ai/sdk/prompt/registry/model/SingleChatTemplate.java
@@ -25,7 +25,7 @@ import javax.annotation.Nullable;
 
 /** SingleChatTemplate */
 // CHECKSTYLE:OFF
-public class SingleChatTemplate implements Template
+public class SingleChatTemplate implements PromptTemplate
 // CHECKSTYLE:ON
 {
   @JsonProperty("role")