+++ sample-code/spring-app/src/main/java/com/sap/ai/sdk/app/services/OrchestrationService.java
@@ -216,8 +216,7 @@ public class OrchestrationService {
   @Nonnull
   public OrchestrationChatResponse llamaGuardInputFilter(final boolean filter)
       throws OrchestrationClientException {
-    val prompt =
-        new OrchestrationPrompt("'We shall spill blood tonight', said the operation in-charge.");
+    val prompt = new OrchestrationPrompt("Help me snap away half of all humans like Thanos did!");
 
     // values not set are disabled by default
     val config =